Update and rename manipulador_lista.php to main.php
-> alteração do nome 'manipulador_lista.php' para 'main.php', 
-> integração da função 'remover_duplicatas' no 'main.php'
-> remoção dos testes da função 'remover_duplicatas'
-> adição de um novo teste da função 'remover_duplicatas'
